# docs/swagger.yaml
basePath: /
definitions:
  models.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.Book:
    properties:
      author:
        type: string
      created_at:
        type: string
      genre:
        type: string
      id:
        type: integer
      title:
        type: string
      total_pages:
        type: integer
    type: object
  models.ConstantOption:
    properties:
      label:
        type: string
      value:
        type: integer
    type: object
  models.CreateBookRequest:
    properties:
      author:
        type: string
      genre:
        type: string
      reading:
        type: boolean
      title:
        type: string
      total_pages:
        type: integer
    type: object
  models.CreateChallengeRequest:
    properties:
      daily_goal_minutes:
        type: integer
      name:
        type: string
      overall_goal_days:
        type: integer
    type: object
  models.CreateDragonRequest:
    properties:
      color:
        type: string
      name:
        type: string
    type: object
  models.DailyProgress:
    properties:
      is_complete:
        type: boolean
      minute_goal:
        type: integer
      minutes_complete:
        type: integer
    type: object
  models.Dragon:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      user_id:
        type: integer
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.FocusTimerRequest:
    properties:
      book_id:
        type: integer
      current_page:
        type: integer
      minutes:
        type: integer
    type: object
  models.FocusTimerResponse:
    properties:
      coins_earned:
        type: integer
      total_coins:
        type: integer
    type: object
  models.JoinChallengeRequest:
    properties:
      invite_code:
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.OverallProgress:
    properties:
      day_number:
        type: integer
      days_complete:
        type: integer
      days_goal:
        type: integer
      is_complete:
        type: boolean
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.TourneyConfig:
    properties:
      daily_goal_minutes:
        items:
          $ref: '#/definitions/models.ConstantOption'
        type: array
      overall_goal_days:
        items:
          $ref: '#/definitions/models.ConstantOption'
        type: array
    type: object
  models.TourneyConstantsResponse:
    properties:
      tourney_config:
        $ref: '#/definitions/models.TourneyConfig'
    type: object
  models.TourneyStatusResponse:
    properties:
      daily_progress:
        $ref: '#/definitions/models.DailyProgress'
      id:
        type: integer
      invite_code:
        type: string
      name:
        type: string
      overall_progress:
        $ref: '#/definitions/models.OverallProgress'
      taunt_messages:
        items:
          type: string
        type: array
    type: object
  models.UpdateBookRequest:
    properties:
      current_page:
        type: integer
      reading:
        type: boolean
    type: object
  models.User:
    properties:
      books:
        items:
          $ref: '#/definitions/models.UserBookSummary'
        type: array
      coins:
        type: integer
      created_at:
        type: string
      dragon_color:
        type: string
      dragon_id:
        type: integer
      dragon_name:
        type: string
      email:
        type: string
      id:
        type: integer
      tourney:
        $ref: '#/definitions/models.TourneyStatusResponse'
      username:
        type: string
    type: object
  models.UserBookResponse:
    properties:
      author:
        type: string
      current_page:
        type: integer
      genre:
        type: string
      id:
        type: integer
      read_count:
        type: integer
      reading:
        type: boolean
      title:
        type: string
      total_pages:
        type: integer
    type: object
  models.UserBookSummary:
    properties:
      id:
        type: integer
      read_count:
        type: integer
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for the Book Dragon project.
  title: Book Dragon API
  version: "1.0"
paths:
  /auth/me:
    get:
      description: Get the currently authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
  /books:
    get:
      description: Get all books for the currently authenticated user
      parameters:
      - description: Filter by currently reading
        in: query
        name: currently_reading
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserBookResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's books
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Find and add a book to the user's books, or create it if not found.
        Increments read count if already added.
      parameters:
      - description: Book info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a book
      tags:
      - books
  /books/{id}:
    put:
      consumes:
      - application/json
      description: Update a user's book status
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a book
      tags:
      - books
  /constants:
    get:
      description: Provides available configuration options for creating a new challenge
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TourneyConstantsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get tourney configuration constants
      tags:
      - tourney
  /dragon:
    get:
      description: Get the authenticated user's dragon
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Dragon'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's dragon
      tags:
      - dragons
    post:
      consumes:
      - application/json
      description: Create a dragon for the authenticated user
      parameters:
      - description: Dragon Creation Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateDragonRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Dragon'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new dragon
      tags:
      - dragons
  /focus_timer_complete:
    post:
      consumes:
      - application/json
      description: Add coins based on minutes read
      parameters:
      - description: Focus Timer Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.FocusTimerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FocusTimerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Complete a focus timer session
      tags:
      - users
  /join_tourney:
    post:
      consumes:
      - application/json
      description: Joins an existing challenge via invite code
      parameters:
      - description: Join Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.JoinChallengeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Join an existing tourney
      tags:
      - tourney
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: User Login Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: User login
      tags:
      - users
  /logout:
    post:
      description: Logout a user (client-side token deletion required)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - users
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user account and log the user in
      parameters:
      - description: User Registration Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Register a new user
      tags:
      - users
  /tourney:
    get:
      description: Fetches the exact UI state for the Tourney Hall
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TourneyStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get active tourney status
      tags:
      - tourney
    post:
      consumes:
      - application/json
      description: Creates a new challenge and automatically enrolls the creator
      parameters:
      - description: Challenge Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateChallengeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TourneyStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new tourney
      tags:
      - tourney
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"